rename leon to svd
